+++ sdk/storage/azure-storage-file-datalake/src/test/java/com/azure/storage/file/datalake/FileAPITest.groovy
@@ -68,6 +68,18 @@ class FileAPITest extends APISpec {
         thrown(StorageErrorException)
     }
 
+    def "Create overwrite"() {
+        when:
+        fc = fsc.getFileClient(generatePathName())
+        fc.create()
+
+        // Try to create the resource again
+        fc.create(false)
+
+        then:
+        thrown(StorageErrorException)
+    }
+
     def "Exists"() {
         when:
         fc = fsc.getFileClient(generatePathName())
@@ -1397,6 +1409,18 @@ class FileAPITest extends APISpec {
 
         when:
         fc.flush(1)
+
+        then:
+        thrown(StorageErrorException)
+    }
+
+    def "Flush data overwrite"() {
+        when:
+        fc.append(new ByteArrayInputStream(defaultData.array()), 0, defaultDataSize)
+        fc.flush(defaultDataSize)
+        fc.append(new ByteArrayInputStream(defaultData.array()), 0, defaultDataSize)
+        // Attempt to write data without overwrite enabled
+        fc.flush(defaultDataSize, false)
 
         then:
         thrown(StorageErrorException)